+++ Ejercicio2/E2V1/iRobot2aParte.cpp
@@ -9,7 +9,7 @@
 
 iRobot2aParte::iRobot2aParte(long type) {
 	this->type = type;
-	colaSalida = new Queue<ColaSalida::message>(IPC::path, (long) IPC::QueueIdentifier::SALIDA_FROM_CTL_TO_INTERFACE, "iRobot2aParte");
+	colaSalida = new Queue<ColaSalida::message>(IPC::path, (int) IPC::QueueIdentifier::SALIDA_FROM_CTL_TO_INTERFACE, "iRobot2aParte");
 	colaSalida->get();
 }
